+++ src/index.ts
@@ -335,6 +335,9 @@ export {
   bookzenPreset,
   BOOKZEN_FIELDS,
   BOOKZEN_TAGS,
+  samforaPreset,
+  SAMFORA_FIELDS,
+  SAMFORA_TAGS,
 } from './vendors';
 
 export type {
@@ -352,4 +355,8 @@ export type {
   BookzenFieldNames,
   BookzenTagSchema,
   BookzenTagNames,
+  SamforaFieldSchema,
+  SamforaFieldNames,
+  SamforaTagSchema,
+  SamforaTagNames,
 } from './vendors';